requirements.txt
aiohttp
feedparser
pydantic
python-dotenv
google-generativeai
python-telegram-bot
beautifulsoup4
